Change scrapy method to parse only
CSS selectors is not working for this specific site, as an alternative, an API call can be made to still scrape the data which is in this case through output through a JSON file. Previous function is excluded entirely as they are currently redundant.

A list comprehension loop is used to get the value of primary data of each job listings. JSON file contains job data with max of 15 lists.